fix(utils): Move some stuff to lib
