+++ src/app/verhuren/page.tsx
@@ -7,7 +7,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { useEffect, useState } from "react";
 import { Controller, useForm } from "react-hook-form";
-import { hasToken } from "../api/auth/Cookies";
+import { hasToken } from "../../lib/cookieUtil";
 import { getImage, postImages } from "../api/Images-controller";
 import { checkStripeConnection, linkToStripe } from "../api/Payment-controller";
 import { postTrailer } from "../api/Trailer-controller";
@@ -470,11 +470,10 @@ const Verhuren = () => {
                 <div
                   key={day}
                   aria-label={day}
-                  className={`flex flex-col items-center justify-center rounded w-14 h-20 cursor-pointer ${
-                    !watch(`availability.${day}`)
-                      ? "bg-primary-100 text-white"
-                      : "bg-offWhite-100"
-                  }`}
+                  className={`flex flex-col items-center justify-center rounded w-14 h-20 cursor-pointer ${!watch(`availability.${day}`)
+                    ? "bg-primary-100 text-white"
+                    : "bg-offWhite-100"
+                    }`}
                   onClick={() => toggleDay(day)}
                 >
                   <p className="font-bold">{getDayAbbreviation(day)}</p>
