+++ src/lib/classNameUtil.ts
@@ -1,0 +1,6 @@
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
+
+export function getMergedTailwindClassNames(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
